requirements.txt
streamlit>=1.35.0
firebase-admin>=6.4.0
pdfplumber>=0.10.3
pandas>=2.2.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
plotly>=5.22.0
